Revert "refactor: LLMClient and ToolManager"
